Use either SAS or WWN id for match, and include Enc.ID
The existing tooling assumed enclosure id# 0 with blink/unblink,
as this may not be the case and scraping the enclosure id is
fairly easy, this patch does so within megamap and uses a string
containing both the enc.id and disk.id in megablink instead of
the hard-coded zero.

Additionally, and this part of the patch is inelegant, I
had an issue where the disks with a certain brand in my array did
not match via SAS Address, but did (close-)match via WWN, while
others would only (close-)match via SAS address.  I extended the
code to scrape both and use whichever matched first (WWN comes
first in the megacli output).